Workaround on imperceptible bug
The current add_inlines() overwrites self.inlines in a way that if called more than once will crash the app.

Calling more than once is as easy as calling get_urls() more than once as that one calls add_inlines.

The dirty hack will simply not instantiate the inline if it was already instantiated. Done this way as it is the least intrusive and this code is currently not tested.